// web/src/api/generated/models/ProfileVisibility.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ProfileVisibility = {
    SELF_PROFILE: 'SELF_PROFILE',
    PRIVATE_PARTNER_NOTE: 'PRIVATE_PARTNER_NOTE'
} as const;
export type ProfileVisibility = typeof ProfileVisibility[keyof typeof ProfileVisibility];


export function instanceOfProfileVisibility(value: any): boolean {
    for (const key in ProfileVisibility) {
        if (Object.prototype.hasOwnProperty.call(ProfileVisibility, key)) {
            if (ProfileVisibility[key as keyof typeof ProfileVisibility] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ProfileVisibilityFromJSON(json: any): ProfileVisibility {
    return ProfileVisibilityFromJSONTyped(json, false);
}

export function ProfileVisibilityFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileVisibility {
    return json as ProfileVisibility;
}

export function ProfileVisibilityToJSON(value?: ProfileVisibility | null): any {
    return value as any;
}

export function ProfileVisibilityToJSONTyped(value: any, ignoreDiscriminator: boolean): ProfileVisibility {
    return value as ProfileVisibility;
}