// web/src/api/generated/models/ProfilePreferenceView.ts
export interface ProfilePreferenceView {
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceView
     */
    accountId: string;
    /**
     * 
     * @type {PreferenceCategory}
     * @memberof ProfilePreferenceView
     */
    category: PreferenceCategory;
    /**
     * 
     * @type {Date}
     * @memberof ProfilePreferenceView
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceView
     */
    id: string;
    /**
     * 
     * @type {PreferenceSentiment}
     * @memberof ProfilePreferenceView
     */
    sentiment: PreferenceSentiment;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceView
     */
    topic: string;
    /**
     * 
     * @type {Date}
     * @memberof ProfilePreferenceView
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceView
     */
    value: string;
    /**
     * 
     * @type {number}
     * @memberof ProfilePreferenceView
     */
    version: number;
    /**
     * 
     * @type {ProfileVisibility}
     * @memberof ProfilePreferenceView
     */
    visibility: ProfileVisibility;
}